fix(webhook REST): refactored controller for webhook REST API
